+++ src/components/datagrid/datagrid.js
@@ -3202,29 +3202,17 @@ Datagrid.prototype = {
         if (col.component) {
           self.tableBody.find('tr').each(function () {
             const row = $(this);
-            const rowIdx = row.attr('data-index');
+            const rowIdx = self.settings.treeGrid ?
+              self.actualPagingRowIndex(self.actualRowIndex(row)) :
+              self.dataRowIndex(row);
             const lineage = row.attr('data-lineage');
-            let value = self.settings.dataset;
-            if (lineage) {
-              const drilldown = lineage.split('.');
-              drilldown.push(rowIdx);
-              let first = true;
-              drilldown.forEach((childIdx) => {
-                if (first && value[childIdx]) {
-                  value = value[childIdx];
-                } else if (value.children && value.children[childIdx]) {
-                  value = value.children[childIdx];
-                }
-                first = false;
-              });
-            } else {
-              value = value[rowIdx];
-            }
+            const rowData = self.rowData(rowIdx);
             const colIdx = self.columnIdxById(col.id);
             const args = {
-              row: rowIdx,
+              row: lineage || rowIdx,
               cell: colIdx,
-              value,
+              value: rowData,
+              rowData,
               col,
               api: self
             };
@@ -8093,7 +8081,11 @@ Datagrid.prototype = {
       this.fieldValue(this.settings.dataset[row], col.field));
 
     if (col.isEditable) {
-      const canEdit = col.isEditable(row, cell, cellValue, col, this.settings.dataset[row]);
+      let rowData = this.settings.dataset[row];
+      if (this.settings.treeDepth && this.settings.treeDepth[row]) {
+        rowData = this.settings.treeDepth[row].node;
+      }
+      const canEdit = col.isEditable(row, cell, cellValue, col, rowData, this, 'is-editable');
 
       if (!canEdit) {
         return false;
@@ -8151,10 +8143,10 @@ Datagrid.prototype = {
     }
 
     const thisRow = this.actualRowNode(row);
-    const idx = this.settings.treeGrid ? this.actualRowIndex(thisRow) : this.dataRowIndex(thisRow);
-    const rowData = this.settings.treeGrid ?
-      this.settings.treeDepth[idx].node :
-      this.settings.dataset[idx];
+    const idx = this.settings.treeGrid ? this.actualPagingRowIndex(this.actualRowIndex(thisRow)) :
+      this.dataRowIndex(thisRow);
+    const rowData = this.rowData(this.dataRowIndex(thisRow));
+
     const cellWidth = cellParent.outerWidth();
     const isEditor = $('.is-editor', cellParent).length > 0;
     const isPlaceholder = $('.is-placeholder', cellNode).length > 0;
@@ -8253,6 +8245,18 @@ Datagrid.prototype = {
     this.element.triggerHandler('entereditmode', [{ row: idx, cell, item: rowData, target: cellNode, value: cellValue, column: col, editor: this.editor }]);
 
     return true;  //eslint-disable-line
+  },
+
+  /**
+   * Get the data for a row node
+   * @private
+   * @param {object} rowNode The jquery row node.
+   * @returns {object} The row of data from the dataset.
+   */
+  rowData(rowIdx) {
+    return this.settings.treeGrid ?
+      this.settings.treeDepth[rowIdx].node :
+      this.settings.dataset[rowIdx];
   },
 
   /**
@@ -9015,9 +9019,7 @@ Datagrid.prototype = {
     if (dataRowIndex === null || dataRowIndex === undefined || isNaN(dataRowIndex)) {
       dataRowIndex = row;
     }
-    const rowData = isTreeGrid ?
-      this.settings.treeDepth[row].node :
-      this.settings.dataset[dataRowIndex];
+    const rowData = this.rowData(dataRowIndex);
 
     if (rowNodes.length === 0 && this.settings.paging) {
       // TODO Frozen Editing with Paging